reafctor: move API to core
