+++ src/app/features/todos/data/infrastructure/local_storage/local.storage.todos.api.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, lastValueFrom, map, Observable, of } from 'rxjs';
-import { TodosApi } from '../../../domain/infrastructure/todos_api';
+import { TodosApi } from '../../../../../core/domain/api/todos_api';
 import { Todo } from '../../../../../core/domain/model/todo.model';
 import { TodoNotFoundException } from 'src/app/core/domain/exceptions/exceptions';
